# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a microservice for expense management.",
        "title": "Squirrel API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
...
        "/transactions": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a list of all transactions for the authenticated app with optional filtering. Stats in response only include expenses.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "List all transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by category ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by type (income, expense)",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by recurring status (0 or 1, default 0)",
                        "name": "recurring",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by predefined date ranges (today, yesterday, this month, last month, this year, last year)",
                        "name": "dated",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter from date (YYYY-MM-DD)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to date (YYYY-MM-DD)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_transaction.TransactionListResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create a new transaction with the provided amount, type, category, recurring status and dated",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Create a new transaction",
                "parameters": [
                    {
                        "description": "Transaction object",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transaction.CreateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_transaction.Transaction"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    }
                }
            }
        },
        "/transactions/stats": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get statistics for transactions for the authenticated app with optional filtering. Only expenses are included.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get transaction statistics",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Filter by category ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filter by recurring status (0 or 1, default 0)",
                        "name": "recurring",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by predefined date ranges (today, yesterday, this month, last month, this year, last year)",
                        "name": "dated",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter from date (YYYY-MM-DD)",
                        "name": "from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter to date (YYYY-MM-DD)",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_transaction.TransactionStats"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    }
                }
            }
        },
        "/transactions/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get a single transaction by its ID if it belongs to the app",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Get transaction by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_transaction.Transaction"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update an existing transaction including amount, type, category, recurring status and dated if it belongs to the app",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Update transaction by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Transaction object",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transaction.UpdateTransactionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_transaction.Transaction"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Delete a transaction by its ID if it belongs to the app",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Delete transaction by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/squirrel_internal_platform_render.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "internal_transaction.CategoryAmountSum": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "integer"
                },
                "total_sum": {
                    "type": "number"
                }
            }
        },
        "internal_transaction.CreateTransactionRequest": {
            "type": "object",
            "required": [
                "amount",
                "type"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category_id": {
                    "type": "integer"
                },
                "dated": {
                    "type": "string"
                },
                "recurring": {
                    "type": "integer"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "income",
                        "expense"
                    ]
                }
            }
        },
        "internal_transaction.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "app_id": {
                    "type": "integer"
                },
                "category_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "dated": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "recurring": {
                    "type": "integer"
                },
                "type": {
                    "description": "income, expense",
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "internal_transaction.TransactionListResponse": {
            "type": "object",
            "properties": {
                "stats": {
                    "$ref": "#/definitions/internal_transaction.TransactionStats"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_transaction.Transaction"
                    }
                }
            }
        },
        "internal_transaction.TransactionStats": {
            "type": "object",
            "properties": {
                "category_top_10_by_amount_sum": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_transaction.CategoryAmountSum"
                    }
                },
                "category_wise_amount_sum": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_transaction.CategoryAmountSum"
                    }
                },
                "top_10_by_amount": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_transaction.Transaction"
                    }
                }
            }
        },
        "internal_transaction.UpdateTransactionRequest": {
            "type": "object",
            "required": [
                "amount",
                "type"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category_id": {
                    "type": "integer"
                },
                "dated": {
                    "type": "string"
                },
                "recurring": {
                    "type": "integer"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "income",
                        "expense"
                    ]
                }
            }
        },
        "squirrel_internal_platform_render.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}